% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmap_track.r
\name{centroidXYdf}
\alias{centroidXYdf}
\title{quick and dirty centroid function.
does not take spherical coords into account.}
\usage{
centroidXYdf(df)
}
\description{
quick and dirty centroid function.
does not take spherical coords into account.
}
