% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.r
\name{cm.cols1}
\alias{cm.cols1}
\alias{cm.cols1.1}
\alias{cm.cols1.2}
\alias{cm.cols.sd}
\alias{cm.cols.dif}
\title{Color palette from Cory}
\usage{
cm.cols1(x, bias = 1)

cm.cols1.1(x, bias = 1)

cm.cols1.2(x, bias = 1)

cm.cols.sd(x, bias = 1)

cm.cols.dif(x, bias = 1)
}
\description{
Color palette from Cory
}
\section{Functions}{
\itemize{
\item \code{cm.cols1}: basic color scheme

\item \code{cm.cols1.1}: more highs and lows

\item \code{cm.cols1.2}: no grey

\item \code{cm.cols.sd}: faded cm.cols1

\item \code{cm.cols.dif}: for plotting differences
}}

