% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfExtent.r
\name{dfExtent}
\alias{dfExtent}
\title{accepts a dataframe (assuming lon/lat columns)
returns a list with values cooresponding to an extent object}
\usage{
dfExtent(dat)
}
\description{
accepts a dataframe (assuming lon/lat columns)
returns a list with values cooresponding to an extent object
}
