% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr.r
\name{escapeForLatex}
\alias{escapeForLatex}
\title{Escapes latex special characters.
Add additional to this function as needed.}
\usage{
escapeForLatex(str)
}
\description{
Escapes latex special characters.
Add additional to this function as needed.
}
