% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorfuns.r
\name{gdaldemColorTable}
\alias{gdaldemColorTable}
\title{Creates a color table that can be used by gdaldem
Write the return to disk using:
write.table(ret,'myfile.txt', row.names=FALSE,quote=FALSE,col.names=FALSE)}
\usage{
gdaldemColorTable(dat)
}
\description{
Creates a color table that can be used by gdaldem
Write the return to disk using:
write.table(ret,'myfile.txt', row.names=FALSE,quote=FALSE,col.names=FALSE)
}
