% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfExtent.r
\name{geeExtentStr}
\alias{geeExtentStr}
\title{accepts a dataframe (assuming lon/lat columns)
returns a string specifying a gee bounding box}
\usage{
geeExtentStr(dat, digits = 7)
}
\arguments{
\item{digits}{\code{integer} Number of digits to round coordinates}
}
\description{
accepts a dataframe (assuming lon/lat columns)
returns a string specifying a gee bounding box
}
