% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmap_track.r
\name{getMapRetry.data.frame}
\alias{getMapRetry.data.frame}
\title{Gets a google map, retries if fails.}
\usage{
\method{getMapRetry}{data.frame}(x, ...)
}
\arguments{
\item{x}{\code{data.frame}. Data frame should have lon/lat columns}
}
\value{
a ggmap object (a classed raster object with a bounding box attribute)
}
\description{
Gets a google map, retries if fails.
}
\examples{
getMapRetry(mydf,11,'Satellite')
}
