% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorfuns.r
\name{tuple2hex}
\alias{tuple2hex}
\title{Accpets rbg tuple string e.g. '165 239 210' and returns hex string '#A5EFD2'}
\usage{
tuple2hex(x)
}
\arguments{
\item{x}{\code{string} vector of rgb tuples in the form 'r g b'.}
}
\description{
Accpets rbg tuple string e.g. '165 239 210' and returns hex string '#A5EFD2'
}
\examples{
tuple2hex(c("165 239 210", "111 45 93"))
}
